backend/requirements.txt
# Pin setuptools so numba/llvmlite (from umap-learn) can build; setuptools>=70 breaks their build
setuptools<70
wheel
fastapi
uvicorn
psycopg2-binary
python-dotenv
pandas
sentence-transformers
numpy
geopandas
supabase
google-genai
scikit-learn
gradium